Implement api

+++ src/main/kotlin/org/bigcraft/lootpool/core/LootPoolManager.kt
@@ -4,15 +4,18 @@ import com.google.inject.Inject
 import com.google.inject.Singleton
 import org.bigcraft.lootpool.AbstractManager
 import org.bigcraft.lootpool.ConfigurableFactory
+import org.bigcraft.lootpool.Load
+import org.bigcraft.lootpool.ReloadConfig
 import org.bigcraft.lootpool.api.Loot
 import org.bigcraft.lootpool.api.LootPool
+import org.bigcraft.lootpool.api.LootPoolProvider
 import org.bukkit.configuration.ConfigurationSection
 import org.bukkit.configuration.file.YamlConfiguration
 import org.bukkit.configuration.serialization.ConfigurationSerialization
 import java.io.File
 
 @Singleton
-class LootPoolManager @Inject constructor(plugin: org.bigcraft.lootpool.LootPool) : AbstractManager<LootPool>() {
+class LootPoolManager @Inject constructor(private val plugin: org.bigcraft.lootpool.LootPool) : AbstractManager<LootPool>(), LootPoolProvider {
 
     override val sectionKey: String = "lootpool"
     override val valueLoader: ConfigurableFactory<LootPool> = LootPoolFactory
@@ -25,25 +28,36 @@ class LootPoolManager @Inject constructor(plugin: org.bigcraft.lootpool.LootPool
         ConfigurationSerialization.registerClass(LootPool::class.java)
     }
 
-    fun getLootPool(key: String) = loadedMap[key]
+    override fun getLootPool(key: String) = loadedMap[key]
 
-    fun removeLootPool(key: String) {
-        loadedMap.remove(key)
+    override fun removeLootPool(key: String): LootPool? {
+        val lootPool = loadedMap.remove(key)
         val file = File(lootDirectory, "$key.yml")
-        file.delete()
+        if (file.exists())
+            file.delete()
+        return lootPool
     }
 
-    override fun load(config: ConfigurationSection) {
-        super.load(config)
-        loadFiles()
+    override fun addLootPool(key: String, lootPool: LootPool) {
+        loadedMap[key] = lootPool
     }
 
-    fun write(key: String, lootPool: LootPool) {
+    override fun writeLootPool(key: String, lootPool: LootPool) {
         loadedMap[key] = lootPool
         val file = File(lootDirectory, "$key.yml")
         val config = YamlConfiguration.loadConfiguration(file)
         config.set(key, lootPool)
         config.save(file)
+    }
+
+    override fun reload() {
+        ReloadConfig.run(plugin)
+        Load.run(plugin)
+    }
+
+    override fun load(config: ConfigurationSection) {
+        super.load(config)
+        loadFiles()
     }
 
     private fun loadFiles() {
